+++ src/components/general/zone-title-card/RainFx.jsx
@@ -2,12 +2,12 @@ import * as THREE from 'three';
 import React, { useState, useEffect, useRef } from 'react';
 // import classnames from 'classnames';
 import {world} from '../../../../world.js';
-import {RainBgFxMesh} from '../../../../background-fx/rain.js';
+import {RainBgFxMesh} from '../../../../background-fx/RainBgFx.js';
 
 import styles from './zone-title-card.module.css';
 
-const localVector4D = new THREE.Vector4();
-const localColor = new THREE.Color();
+// const localVector4D = new THREE.Vector4();
+// const localColor = new THREE.Color();
 
 //
